channel_hash: widen substrate-wide u16 → u32
Substrate-wide widening of the canonical channel-hash type from u16
to u32. The wire `NetHeader::channel_hash` stays u16 per the
`origin_hash u64-canonical / u32-wire` precedent — per-packet width
is fixed by the 64-byte header budget, and wire-side bucket
collisions are benign (only affect routing fast-path filter
precision, not ACL or storage decisions, which now key on the wider
canonical hash).

- New `pub type ChannelHash = u32` (canonical) alongside the existing
  wire `u16`; `ChannelName::hash()` returns canonical, new
  `ChannelName::wire_hash()` returns the truncated wire value.
- `ChannelConfigRegistry` gains `by_wire_hash` (u16 → names) for
  receive-side dispatch, keeping `by_hash` keyed on u32 canonical.
- `AuthGuard`, `PermissionToken`, `TokenCache`, `TokenScope` migrate
  to canonical u32; bloom-filter key buffer widens to 12 bytes.
  Token wire format grows from 159 → 161 bytes (channel_hash field
  2 → 4 bytes).
- `MeshNode::register_rpc_inbound` accepts canonical `ChannelHash`;
  the inbound dispatcher map indexes by wire u16 but stores
  `(canonical, dispatcher)` pairs so wire-bucket collisions don't
  share a dispatcher slot, and `RpcInboundEvent::channel_hash` is
  the canonical u32.
- FFI: `net_channel_hash` widens `*mut u16` → `*mut u32`; Python /
  Node / Go bindings widen their `channel_hash` exports
  correspondingly.
- Greedy data-plane observe path stays on wire u16 (only signal
  available at packet decode time); the cache file is named
  `dataforts/greedy/<hex16>` as before.

+++ net/crates/net/include/net.go.h
@@ -703,8 +703,10 @@ int      net_mesh_set_reflex_override(net_meshnode_t* handle,
  * active. */
 int      net_mesh_clear_reflex_override(net_meshnode_t* handle);
 
-/* Hash a channel name to its 16-bit wire representation. */
-int      net_channel_hash(const char* channel, uint16_t* out_hash);
+/* Hash a channel name to its canonical 32-bit substrate identifier
+ * (used for ACL/storage/config keys; the wire NetHeader fast-path
+ * hint is the low 16 bits of this value). */
+int      net_channel_hash(const char* channel, uint32_t* out_hash);
 
 /* =========================================================================
  * Capabilities (announce / find_nodes).
